ocr-worker/requirements.txt
# OCR
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.26.4
imutils==0.5.4

# Queue
redis==5.0.4
rq==1.16.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

# Storage
boto3==1.34.84

# HTTP
httpx==0.27.0
fastapi==0.110.2
uvicorn==0.29.0

# Config / Utils
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1

